fix loop-fusion bugs
There were misstype bugs that allowed involuntarily to validate the dominance check and
loop adjacency check.
The previous check should fail in case of loop in rotated form with guards.
But to handle the case of a loop from 0 to N where N's value is determined only at runtime,
in case of loop rotated forms with guards the check of adjacency and dominance are not done.

Signed-off-by: jot <singhamanjot726@gmail.com>